fix: TrackResult expects Float values, not Int
Changed pan/tilt from Int to Float in ObjectTracker.kt:
- TrackResult.panDelta and tiltDelta are Float type
- Was incorrectly converting to Int with .toInt()
- Now passes Float values directly
- Maintains sensitivity application and clamping

+++ app/src/main/java/org/rhanet/roverctrl/tracking/ObjectTracker.kt
@@ -203,8 +203,9 @@ class ObjectTracker(
         // Apply sensitivity multipliers and clamp to [-100, 100]
         val panFloat  = (panRaw * panSensitivity).coerceIn(-100f, 100f)
         val tiltFloat = (tiltRaw * tiltSensitivity).coerceIn(-100f, 100f)
-        val pan  = panFloat.toInt()
-        val tilt = tiltFloat.toInt()
+        // TrackResult expects Float values, not Int
+        val pan  = panFloat
+        val tilt = tiltFloat
 
         val mode = if (shouldDetect) "DETECT" else "TRACK"
         Log.d(TAG, "[$mode] ${finalDetection.label} ${(finalDetection.confidence*100).toInt()}% " +
